flashlight finally rendering and moving with player (kind of...)
