+++ SummativeGame/src/entities/mobs/Flashlight.java
@@ -2,11 +2,10 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package graphics;
+package entities.mobs;
 
 import coreEngine.Game;
 import coreEngine.GameVariables;
-import entities.Entity;
 import java.awt.Graphics;
 import java.awt.image.BufferedImage;
 
@@ -14,12 +13,12 @@ import java.awt.image.BufferedImage;
  *
  * @author pistm9061
  */
-public class Flashlight extends Entity{
+public class Flashlight extends Mob{
     
     BufferedImage flashlight;
     private Game game;
     
-    public Flashlight(Game game, BufferedImage flashlight, float x, float y, float width, float height){
+    public Flashlight(Game game, BufferedImage flashlight, float x, float y){
         super(x, y, GameVariables.getFlashlightDiameter(), GameVariables.getFlashlightDiameter());
         this.flashlight = flashlight;
     }
@@ -33,4 +32,10 @@ public class Flashlight extends Entity{
     public void render(Graphics g) {
         g.drawImage(flashlight, (int) x, (int) y, width, height, null);
     }
+    
+    public void move(float dx, float dy){
+		//moves the player the passed in x and y displacements
+		super.move(dx, dy);		
+	}
+    
 }
